front: extract timetable train list in its own component
This will be usefull to display a list of trains in the package mode
tabs.
