+++ front/src/applications/operationalStudies/views/Scenario/components/Timetable/Timetable.tsx
@@ -1,35 +1,16 @@
-import { useMemo, useCallback, useState } from 'react';
+import { useState } from 'react';
 
 import cx from 'classnames';
-import { useSelector } from 'react-redux';
-import { Virtualizer } from 'virtua';
 
-import { useScenarioContext } from 'applications/operationalStudies/hooks/useScenarioContext';
-import { MANAGE_TIMETABLE_ITEM_TYPES } from 'applications/operationalStudies/views/Scenario/consts';
-import { useSubCategoryContext } from 'common/SubCategoryContext';
-import type {
-  TimetableItemWithDetails,
-  PacedTrainWithDetails,
-  TrainScheduleWithDetails,
-} from 'modules/timetableItem/types';
-import { selectTrainToEdit } from 'reducers/osrdconf/operationalStudiesConf';
+import type { TimetableItemWithDetails } from 'modules/timetableItem/types';
 import type {
   TimetableItemId,
   TimetableItem,
-  OccurrenceId,
   TimetableItemToEditData,
 } from 'reducers/osrdconf/types';
-import {
-  getSelectedTrainId,
-  getTrainIdUsedForProjection,
-} from 'reducers/simulationResults/selectors';
-import { useAppDispatch } from 'store';
-import { useDateTimeLocale } from 'utils/date';
-import { isPacedTrainWithDetails, isTrainScheduleId } from 'utils/trainId';
 
-import PacedTrainItem from './PacedTrain/PacedTrainItem';
+import CalendarTrainList from './CalendarTrainList';
 import TimetableToolbar from './TimetableToolbar';
-import TrainScheduleItem from './TrainScheduleItem';
 import useFilterTimetableItems from './useFilterTimetableItems';
 
 type TimetableProps = {
@@ -46,9 +27,6 @@ type TimetableProps = {
   projectingOnSimulatedPathException: boolean | undefined;
 };
 
-const formatDepartureDate = (d: Date, locale: Intl.Locale) =>
-  d.toLocaleDateString(locale, { weekday: 'long', day: 'numeric', month: 'long', year: 'numeric' });
-
 const Timetable = ({
   setDisplayTimetableItemManagement,
   upsertTimetableItems,
@@ -62,92 +40,12 @@ const Timetable = ({
   selectedTimetableItemIds,
   projectingOnSimulatedPathException,
 }: TimetableProps) => {
-  const dateTimeLocale = useDateTimeLocale();
-
-  const { workerStatus } = useScenarioContext();
-
-  const [expandedTimetableItemIds, setExpandedTimetableItemIds] = useState<Set<TimetableItemId>>(
-    new Set()
-  );
-
   const [showTrainDetails, setShowTrainDetails] = useState(false);
   const [isSelectMode, setIsSelectMode] = useState(false);
   const [timetableMode, setTimetableMode] = useState<'calendar' | 'package'>('calendar');
 
-  const selectedTrainId = useSelector(getSelectedTrainId);
-  const trainIdUsedForProjection = useSelector(getTrainIdUsedForProjection);
-  const dispatch = useAppDispatch();
-
   const { filteredTimetableItems, ...timetableFilters } =
     useFilterTimetableItems(timetableItemsWithDetails);
-
-  const handleSelectTimetableItem = useCallback(
-    (id: TimetableItemId) => {
-      const currentSelectedTrainIds: TimetableItemId[] = selectedTimetableItemIds;
-      const index = currentSelectedTrainIds.indexOf(id);
-
-      if (index === -1) {
-        currentSelectedTrainIds.push(id);
-      } else {
-        currentSelectedTrainIds.splice(index, 1);
-      }
-
-      setSelectedTimetableItemIds([...currentSelectedTrainIds]);
-    },
-    [selectedTimetableItemIds]
-  );
-
-  const handleExpandTimetableItem = useCallback((id: TimetableItemId) => {
-    setExpandedTimetableItemIds((prevExpandedIds) => {
-      const newExpandedIds = new Set(prevExpandedIds);
-      if (newExpandedIds.has(id)) {
-        newExpandedIds.delete(id);
-      } else {
-        newExpandedIds.add(id);
-      }
-      return newExpandedIds;
-    });
-  }, []);
-
-  const currentDepartureDates = useMemo(
-    () =>
-      filteredTimetableItems.map((train) => formatDepartureDate(train.startTime, dateTimeLocale)),
-    [filteredTimetableItems, dateTimeLocale]
-  );
-
-  const showDepartureDates = useMemo(() => {
-    let previousDepartureDate = '';
-    return currentDepartureDates.map((date) => {
-      const show = date !== previousDepartureDate;
-      if (show) previousDepartureDate = date;
-      return show;
-    });
-  }, [currentDepartureDates]);
-
-  const selectTimetableItemToEdit = useCallback(
-    (
-      itemToEdit: TimetableItemWithDetails,
-      originalPacedTrain?: PacedTrainWithDetails,
-      occurrenceId?: OccurrenceId
-    ) => {
-      dispatch(selectTrainToEdit({ item: itemToEdit, isOccurrence: !!occurrenceId }));
-      const editData = isPacedTrainWithDetails(itemToEdit)
-        ? {
-            timetableItemId: itemToEdit.id,
-            // param originalPacedTrain is defined only when editing an occurrence
-            originalPacedTrain: originalPacedTrain ?? itemToEdit,
-            occurrenceId,
-          }
-        : {
-            timetableItemId: itemToEdit.id,
-          };
-      setTimetableItemToEditData(editData);
-      setDisplayTimetableItemManagement(MANAGE_TIMETABLE_ITEM_TYPES.edit);
-    },
-    []
-  );
-
-  const subCategories = useSubCategoryContext();
 
   return (
     <div className="scenario-timetable">
@@ -172,53 +70,18 @@ const Timetable = ({
           timetableMode={timetableMode}
           setTimetableMode={setTimetableMode}
         />
-        <Virtualizer overscan={15}>
-          {filteredTimetableItems.map((timetableItem, index) => (
-            <div key={`timetable-train-card-${timetableItem.id}`} data-train-id={timetableItem.id}>
-              {showDepartureDates[index] && (
-                <div className="scenario-timetable-departure-date">
-                  {currentDepartureDates[index]}
-                </div>
-              )}
-              {isTrainScheduleId(timetableItem.id) ? (
-                <TrainScheduleItem
-                  isInSelection={selectedTimetableItemIds.includes(timetableItem.id)}
-                  handleSelectTrain={handleSelectTimetableItem}
-                  train={timetableItem as TrainScheduleWithDetails}
-                  isSelected={workerStatus === 'READY' && selectedTrainId === timetableItem.id}
-                  isModified={timetableItem.id === timetableItemToEditData?.timetableItemId}
-                  upsertTrainSchedules={upsertTimetableItems}
-                  removeTrains={removeAndUnselectTrains}
-                  selectTrainToEdit={selectTimetableItemToEdit}
-                  setSelectedTimetableItemIds={setSelectedTimetableItemIds}
-                  projectionPathIsUsed={
-                    workerStatus === 'READY' && trainIdUsedForProjection === timetableItem.id
-                  }
-                  subCategories={subCategories}
-                  isSelectMode={isSelectMode}
-                />
-              ) : (
-                <PacedTrainItem
-                  pacedTrain={timetableItem as PacedTrainWithDetails}
-                  isInSelection={selectedTimetableItemIds.includes(timetableItem.id)}
-                  selectPacedTrainToEdit={selectTimetableItemToEdit}
-                  handleSelectPacedTrain={handleSelectTimetableItem}
-                  isOccurrencesListOpen={expandedTimetableItemIds.has(timetableItem.id)}
-                  handleOpenOccurrencesList={handleExpandTimetableItem}
-                  isOnEdit={timetableItem.id === timetableItemToEditData?.timetableItemId}
-                  selectedTrainId={selectedTrainId}
-                  upsertTimetableItems={upsertTimetableItems}
-                  removePacedTrains={removeAndUnselectTrains}
-                  setSelectedTimetableItemIds={setSelectedTimetableItemIds}
-                  infraIsCached={workerStatus === 'READY'}
-                  subCategories={subCategories}
-                  projectingOnSimulatedPathException={projectingOnSimulatedPathException}
-                  isSelectMode={isSelectMode}
-                />
-              )}
-            </div>
-          ))}
-        </Virtualizer>
+        <CalendarTrainList
+          setDisplayTimetableItemManagement={setDisplayTimetableItemManagement}
+          upsertTimetableItems={upsertTimetableItems}
+          setTimetableItemToEditData={setTimetableItemToEditData}
+          setSelectedTimetableItemIds={setSelectedTimetableItemIds}
+          removeAndUnselectTrains={removeAndUnselectTrains}
+          timetableItemToEditData={timetableItemToEditData}
+          timetableItemsWithDetails={filteredTimetableItems}
+          selectedTimetableItemIds={selectedTimetableItemIds}
+          projectingOnSimulatedPathException={projectingOnSimulatedPathException}
+          isSelectMode={isSelectMode}
+        />
       </div>
     </div>
   );
